requirements.txt
# Core Streamlit and Data Processing
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.24.0

# Excel and Word Document Processing
openpyxl>=3.1.2
python-docx>=1.1.0
docxtpl>=0.16.0
xlsxwriter>=3.1.0
pypandoc>=1.11

# Date and Time Handling
python-dateutil>=2.8.0

# Template Engine
Jinja2>=3.1.0

# Additional Utilities
Pillow>=10.0.0
requests>=2.31.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# For future AI features (optional but recommended)
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# plotly>=5.15.0 